Make: Fix default link rule for --as-needed
microcom does not compile with LDFLAGS="-Wl,--as-needed" because of a
default make rule with LDFLAGS before all object files. The linker seems
to parse the commandline only once, which results in this case in
undefined references.

This patch defines a custom link rule with LDFLAGS at the end of the
command line.

Signed-off-by: Markus Pargmann <mpa@pengutronix.de>