don't add .gitignore to archives created by git-archive
The contents of a microcom.tar.gz or microcom.zip are not expected to be
readded to a git repository. So don't add the git specific files to such
an archive.

The motivation for this change is that lintian (a static checker for
Debian packages) criticizes that there is no copyright declared for
.gitignore. As .gitignore is hardly copyrightable simply not including
it in the original source is the simplest way to please lintian.