Remove in-tree copy of git changelog
I see no sense in having a changelog file that just contains some
reformating of the complete git history. AFAIK the only place it is/was
used is to include it as /usr/share/doc/microcom/changelog.gz in the
microcom Debian package which is not worth the semi-manual effort to
hatch it.

This also makes the script to generate the changelog file obsolete, so
this is removed, too.