serial: configure CLOCAL/CREAD for non-Linux compatibility
While Linux configures both CLOCAL and CREAD on UARTs by default[1],
other *nix systems might not.
E.g. macOS only sets CREAD, but not CLOCAL with the result that
write(2) indefinitely returns with errno = EAGAIN, waiting for a DCD
activation that never happens.

Fix this by explicitly OR'ing CLOCAL. On Linux, this doesn't
matter, but on macOS, it enables microcom to output user input.

CREAD is added as well to account for systems that don't set it by
default. This is in-line with what Michael R. Sweet's
"Serial Programming Guide for POSIX Operating Systems"[1] suggests.

Patch tested on macOS 10.14.1 and Linux 4.20.13.

[1]: https://github.com/torvalds/linux/blob/v4.20/drivers/tty/serial/serial_core.c#L2513
[2]: https://www.cmrr.umn.edu/~strupp/serial.html

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>