can: Get rid of pthread for can transfers
Since there are ios specific read and write callbacks the needed overhead
for can transfers can be implemented in these which makes the separate
thread superfluous. This simplifies the can driver considerably.

As a consequence microcom doesn't need pthreads any more and the
respective changes in configure.ac can go away, too. The configuration
for lgtm which assert that the AX_PTHREAD macro is available can be
dropped and the travis configuration can be simplified accordingly.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Closes: https://github.com/pengutronix/microcom/issues/8