commands: allow specifying a process to log to
While microcom's stdout can be piped to another processes, it makes
input somewhat of a hassle. Address this by teaching the --logfile
option and the log command that a filename starting with a pipe
character refers to an executable to start and pipe the input into.
The child process shares stderr, but has stdout closed.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>