Don't call async-signal-unsafe free in signal handler
We use malloc elsewhere in the code with signals unmasked, but use free
in a signal handler. This could result in a dead lock when terminating
microcom by signal. Simply drop the free as exit is imminent.

The duplication of exit(0) call is ok, as the call site in the signal
handler will be replaced in a follow-up commit.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Signed-off-by: Jonas Rebmann <jre@pengutronix.de>