microcom: make microcom_exit async-signal-safer
Both printf and exit aren't async signal safe, replace them with write
and _Exit respectively.

The access to ios and ios->exit and callees is still illegal, but at
least we don't run risk reentering stdio and causing a deadlock anymore.

As exit flushes stdio buffers as well, we now risk losing unflushed
stdio output. This is acceptable as we aren't buffering the serial
port output and the program is being terminated abnormally anyway.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>