serial: use flock instead of lock files
Some distributions like Fedora don't allow users access to /var/lock and
instead want applications to use lockdev, which in turn isn't available
in distributions like Debian, as those rightfully state that there is a
much more convenient and and easier to get right interface available in
the form of flock(2).

Switch microcom to use flock instead of the hand rolled lockfiles, which
also avoids the issue of stale lock files, as the lock is automatically
removed when the application closes the serial fd.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>