mux: Ignore .read() returning errors EAGAIN and EWOULDBLOCK
For the telnet backend the fd might be ready to read, but there is only
out-of-band data available.

Up to now this out-of-band data is handled in mux.c, but to prepare
handling all the telnet specific stuff in telnet.c, there must be a way
to let telnet's read function handle only out-of-band data.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>