mux: Handle parsing subnegotiations including IACs
The IACs in a subnegotiation are quoted, too. So unquote while parsing.

If the remote rfc2217 side sends (raw)

	SB COM_PORT_CONTROL NOTIFY_MODEMSTATE_SC 0xff 0xff IAC SE

the double 0xff is to be interpreted as a single 0xff (as IAC = 0xff).

Instead the second 0xff was interpreted as quoting the following IAC and
so the end of the subnegotiation is missed making microcom exit with

	Incomplete SB string

Reported-by: James Puderer <jhpuderer@gmail.com>
Fixes: 1334a93b2718 ("implement telnet rfc2217 mode (experimental), update help")
Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>