telnet: Don't send DO COM_PORT_CONTROL
That request isn't sensible because microcom doesn't implement
to be an UART access server.

Both ser2net and Moxa UART servers reply with WONT COM_PORT_CONTROL to a
DO, so the request is refused. To actually use the COM_PORT_CONTROL
extension only a "client side" WILL (by microcom) and a "server side" DO
(by the remote side) is necessary (and sensible).

The rfc2217 document also reads:

    Typically a client will use WILL and WONT, while an access server
    will use DO and DONT.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>